Update and rename fitting_the_phi--ver1--ECPL--tests.py to fitting_the_phi--ECPL--plots.py
optimizing the script (the loop over the parameter space)
